VS/requirements.txt
# Core Flask Framework
Flask>=3.1.2
Flask-SQLAlchemy>=3.1.1
Flask-JWT-Extended>=4.7.1
Flask-CORS>=6.0.1
Flask-Marshmallow>=1.3.0
Flask-WTF>=1.2.2

# Database drivers (MySQL only)
mysql-connector-python>=9.4.0
PyMySQL>=1.1.2

# Security and authentication
Werkzeug>=3.1.3
cryptography>=46.0.1

# Document processing
python-docx>=1.2.0
PyPDF>=6.0.0
PyMuPDF>=1.26.4

# Database management
SQLAlchemy>=2.0.43
alembic>=1.16.5

# Data serialization
marshmallow>=4.0.1
marshmallow-sqlalchemy>=1.4.2

# Development (optional)
python-dotenv>=1.0.0